refactor(website): rename channel preview file
